basic memory functions moved to own source and header
